Adding Home View Service and renaming views
