+++ src/views/ContactProfile/contactprofile.vue
@@ -1,0 +1,16 @@
+<template>
+    <h1>Contato</h1>
+</template>
+  
+<script lang="ts">
+import { defineComponent } from 'vue'
+
+export default defineComponent({
+  name: 'ContactView'
+})
+</script>
+  
+<style scoped>
+  
+</style>
+
