--- src/views/UserProfile.vue
@@ -1,12 +1,0 @@
-<template>
-    <h1>Teste</h1>
-</template>
-  
-<script lang="ts">
-
-</script>
-  
-<style scoped>
-  
-</style>
-
